Extract CP866 to its own file
